requirements.txt
telethon==1.36.0
cryptg==0.4.0
python-socks[asyncio]==2.4.3
python-dotenv==1.0.0
aiohttp==3.9.1

# Database support
aiosqlite==0.19.0

# Subtitle support
requests==2.31.0

# Web Dashboard
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
websockets==12.0